requirements/base.txt
# Django Framework
# To run backend
django==2.2.7

# Creating rest api
djangorestframework==3.10.3


# MYSQL Client
# To connect with database
mysqlclient==1.4.5


# Django-Extensions
# For additional features
django-extensions==2.2.5


# Formatting shell_plus
Pygments==2.4.2
IPython==7.9.0

pytz==2019.3
six==1.13.0
sqlparse==0.3.0
